fix: skip checkouts table creation if already exists in migration

+++ database/migrations/2024_07_19_161253_create_checkouts_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('checkouts', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->boolean('isFreeze')->default(false);
-            $table->json('billing')->nullable();
-            $table->json('shipping')->nullable();
-            $table->integer('total')->default(0)->nullable();
-            $table->string('paymentType')->nullable();
-            $table->json('extra')->nullable();
-            $table->timestamps();
-            $table->foreign('user_id')->references('ID')->on('wp_users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('checkouts')) {
+            Schema::create('checkouts', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->boolean('isFreeze')->default(false);
+                $table->json('billing')->nullable();
+                $table->json('shipping')->nullable();
+                $table->integer('total')->default(0);
+                $table->string('paymentType')->nullable();
+                $table->json('extra')->nullable();
+                $table->timestamps();
+                $table->foreign('user_id')->references('ID')->on('wp_users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
